# swagger.yaml
openapi: 3.0.0
info:
  title: FULLSTACK JOCESTECH
  contact: {}
  version: '1.0'
servers:
  - url: https://jocestech-backend-fullstack.onrender.com/api/v1
    variables: {}
paths:
...
  /auth/loginAdmin:
    post:
      tags:
        - auth
      summary: login admin
      operationId: loginadmin
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/loginadminrequest'
                - example:
                    email: test@gmail.com
                    password: testPassword
            example:
              email: test@gmail.com
              password: testPassword
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
...
  /users/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: the users id
    get:
      tags:
        - user
      summary: get single user
      operationId: getsingleuser
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
  /user/updateUser:
    patch:
      tags:
        - user
      summary: update user
      operationId: updateuser
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/updateuserrequest'
                - example:
                    name: name smith
                    email: name@gmail.com
            example:
              name: name smith
              email: name@gmail.com
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /user/updateUserPassword:
    patch:
      tags:
        - user
      summary: update user password
      operationId: updateuserpassword
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/updateuserpasswordrequest'
                - example:
                    oldPassword: oldPassword
                    newPassword: newPassword
            example:
              oldPassword: oldPassword
              newPassword: newPassword
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
  /reviews:
    get:
      tags:
        - reviews
      summary: get all reviews
      operationId: getallreviews
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    post:
      tags:
        - reviews
      summary: create review
      operationId: createreview
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/createreviewrequest'
                - example:
                    product: 664dc31c17780a551a945989
                    rating: 5
                    title: i love is product
                    comment: this is the best product
            example:
              product: 664dc31c17780a551a945989
              rating: 5
              title: i love is product
              comment: this is the best product
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /reviews/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: the reviews id
    get:
      tags:
        - reviews
      summary: get single review
      operationId: getsinglereview
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    patch:
      tags:
        - reviews
      summary: update reviews
      operationId: updatereviews
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/updatereviewsrequest'
                - example:
                    product: 664dc31c17780a551a945989
                    rating: 4
                    title: i love is product
                    comment: this is the best product
            example:
              product: 664dc31c17780a551a945989
              rating: 4
              title: i love is product
              comment: this is the best product
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    delete:
      tags:
        - reviews
      summary: delete reviews
      operationId: deletereviews
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /reviews/user/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: the reviews user id
    get:
      tags:
        - reviews
      summary: get single user review
      operationId: getsingleuserreview
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
components:
  schemas:
...
    loginadminrequest:
      title: loginadminrequest
      required:
        - email
        - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      example:
        email: test@gmail.com
        password: secret
    updateuserrequest:
      title: updateuserrequest
      required:
        - name
        - email
      type: object
      properties:
        name:
          type: string
        email:
          type: string
      example:
        name: test
        email: test@gmail.com
    updateuserpasswordrequest:
      title: updateuserpasswordrequest
      required:
        - oldPassword
        - newPassword
      type: object
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
      example:
        oldPassword: newSecret
        newPassword: secret
...
    createreviewrequest:
      title: createreviewrequest
      required:
        - product
        - rating
        - title
        - comment
      type: object
      properties:
        product:
          type: string
        rating:
          type: integer
          format: int32
        title:
          type: string
        comment:
          type: string
      example:
        product: 664dc31c17780a551a945989
        rating: 5
        title: i love is product
        comment: this is the best product
    updatereviewsrequest:
      title: updatereviewsrequest
      required:
        - product
        - rating
        - title
        - comment
      type: object
      properties:
        product:
          type: string
        rating:
          type: integer
          format: int32
        title:
          type: string
        comment:
          type: string
      example:
        product: 664dc31c17780a551a945989
        rating: 4
        title: i love is product
        comment: this is the best product
...